feat: InstanceContainer와 프로퍼티 의존성 주입을 수행해주는 Activity를 생성한다.
