+++ app/src/main/java/woowacourse/shopping/di/activity/ActivityInstanceContainer.kt
@@ -3,17 +3,29 @@ package woowacourse.shopping.di.activity
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
 import com.boogiwoogi.di.Qualifier
-import com.boogiwoogi.di.version2.DiContainer
 import com.boogiwoogi.di.version2.Instance
+import com.boogiwoogi.di.version2.InstanceContainer
 import kotlin.reflect.KClass
 import kotlin.reflect.KParameter
 import kotlin.reflect.full.findAnnotation
 import kotlin.reflect.full.hasAnnotation
 import kotlin.reflect.jvm.jvmErasure
 
-class ActivityDiContainer : DiContainer, DefaultLifecycleObserver {
+class ActivityInstanceContainer : InstanceContainer, DefaultLifecycleObserver {
 
     override var value: MutableList<Instance<out Any>>? = mutableListOf()
+
+    override fun onCreate(owner: LifecycleOwner) {
+        super.onCreate(owner)
+
+        value = mutableListOf()
+    }
+
+    override fun onDestroy(owner: LifecycleOwner) {
+        value = null
+
+        super.onDestroy(owner)
+    }
 
     override fun add(instance: Instance<*>) {
         value?.add(instance)
@@ -32,17 +44,5 @@ class ActivityDiContainer : DiContainer, DefaultLifecycleObserver {
             true -> find(parameter.findAnnotation<Qualifier>()?.simpleName)
             false -> find(parameter.type.jvmErasure)
         }
-    }
-
-    override fun onCreate(owner: LifecycleOwner) {
-        super.onCreate(owner)
-
-        value = mutableListOf()
-    }
-
-    override fun onDestroy(owner: LifecycleOwner) {
-        super.onDestroy(owner)
-
-        value = null
     }
 }
