+++ app/src/main/java/woowacourse/shopping/di/activity/DiActivity.kt
@@ -2,9 +2,15 @@ package woowacourse.shopping.di.activity
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.boogiwoogi.di.version2.DiComponent
+import com.boogiwoogi.di.version2.InstanceContainer
 import woowacourse.shopping.ShoppingApplication
 
-open class DiActivity : AppCompatActivity() {
+open class DiActivity : AppCompatActivity(), DiComponent {
+
+    override val instanceContainer: InstanceContainer = ActivityInstanceContainer().apply {
+        lifecycle.addObserver(this)
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -14,10 +20,8 @@ open class DiActivity : AppCompatActivity() {
 
     private fun setupInjector() {
         ShoppingApplication.injector.inject(
-            container = ActivityDiContainer().apply {
-                lifecycle.addObserver(this)
-            },
-            modules = ActivityModule(this),
+            container = instanceContainer,
+            modules = ActivityModules(this),
             target = this
         )
     }
