+++ woogi-di/src/main/java/com/boogiwoogi/di/version2/DiComponent.kt
@@ -1,0 +1,6 @@
+package com.boogiwoogi.di.version2
+
+interface DiComponent {
+
+    val instanceContainer: InstanceContainer
+}
